Don't build pkg/sdn/plugin on non-Linux
