+++ pkg/cmd/server/kubernetes/node/node_config.go
@@ -43,7 +43,6 @@ import (
 	"github.com/openshift/origin/pkg/dns"
 	"github.com/openshift/origin/pkg/sdn"
 	sdnapi "github.com/openshift/origin/pkg/sdn/apis/network"
-	sdnplugin "github.com/openshift/origin/pkg/sdn/plugin"
 )
 
 // NodeConfig represents the required parameters to start the OpenShift node
@@ -81,9 +80,9 @@ type NodeConfig struct {
 	DNSServer *dns.Server
 
 	// SDNPlugin is an optional SDN plugin
-	SDNPlugin *sdnplugin.OsdnNode
+	SDNPlugin sdn.NodeInterface
 	// SDNProxy is an optional service endpoints filterer
-	SDNProxy *sdnplugin.OsdnProxy
+	SDNProxy sdn.ProxyInterface
 }
 
 func BuildKubernetesNodeConfig(options configapi.NodeConfig, enableProxy, enableDNS bool) (*NodeConfig, error) {
@@ -235,23 +234,14 @@ func BuildKubernetesNodeConfig(options configapi.NodeConfig, enableProxy, enable
 	internalKubeInformers := kinternalinformers.NewSharedInformerFactory(kubeClient, proxyconfig.ConfigSyncPeriod.Duration)
 
 	// Initialize SDN before building kubelet config so it can modify option
-	sdnNodeConfig := &sdnplugin.OsdnNodeConfig{
-		PluginName:         options.NetworkConfig.NetworkPluginName,
-		Hostname:           options.NodeName,
-		SelfIP:             options.NodeIP,
-		RuntimeEndpoint:    options.DockerConfig.DockerShimSocket,
-		MTU:                options.NetworkConfig.MTU,
-		OSClient:           originClient,
-		KClient:            kubeClient,
-		KubeInformers:      internalKubeInformers,
-		IPTablesSyncPeriod: proxyconfig.IPTables.SyncPeriod.Duration,
-		ProxyMode:          proxyconfig.Mode,
-	}
-	sdnPlugin, err := sdnplugin.NewNodePlugin(sdnNodeConfig)
-	if err != nil {
-		return nil, fmt.Errorf("SDN initialization failed: %v", err)
-	}
-	if sdnPlugin != nil {
+	var sdnPlugin sdn.NodeInterface
+	var sdnProxy sdn.ProxyInterface
+	if sdn.IsOpenShiftNetworkPlugin(options.NetworkConfig.NetworkPluginName) {
+		sdnPlugin, sdnProxy, err = NewSDNInterfaces(options, originClient, kubeClient, internalKubeInformers, proxyconfig)
+		if err != nil {
+			return nil, fmt.Errorf("SDN initialization failed: %v", err)
+		}
+
 		// SDN plugin pod setup/teardown is implemented as a CNI plugin
 		server.NetworkPluginName = kubeletcni.CNIPluginName
 		server.NetworkPluginDir = kubeletcni.DefaultNetDir
@@ -302,11 +292,6 @@ func BuildKubernetesNodeConfig(options configapi.NodeConfig, enableProxy, enable
 		}),
 		CertFile: options.ServingInfo.ServerCert.CertFile,
 		KeyFile:  options.ServingInfo.ServerCert.KeyFile,
-	}
-
-	sdnProxy, err := sdnplugin.NewProxyPlugin(options.NetworkConfig.NetworkPluginName, originClient, kubeClient)
-	if err != nil {
-		return nil, fmt.Errorf("SDN proxy initialization failed: %v", err)
 	}
 
 	config := &NodeConfig{
